# centinela-agents/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Centinela · API de agentes antifraude",
    "description": "API de agentes de **Centinela**, plataforma antifraude multimodal.\n\nDos procesos, un mismo contrato de salida (`Decision`):\n\n- `POST /api/v1/transactions/evaluate` — aprobar · validación adicional · bloquear\n- `POST /api/v1/documents/validate` — auténtico · sospechoso · falso\n\nToda decisión es explicable y trazable bajo un `trace_id`. Los casos bloqueados\ny las excepciones siempre pasan por una persona.\n\n**Datos sintéticos.** Este servicio se construyó como proyecto académico y no\ndebe procesar datos reales de clientes sin gobierno y consentimiento.\n",
    "contact": {
      "name": "Grupo Centinela"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/transactions/evaluate": {
      "post": {
        "tags": [
          "transacciones"
        ],
        "summary": "Evalúa una transacción y devuelve una decisión explicable",
        "operationId": "evaluate_api_v1_transactions_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Decision"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/evaluate/batch": {
      "post": {
        "tags": [
          "transacciones"
        ],
        "summary": "Evalúa un lote de transacciones (replay y modo sombra)",
        "operationId": "evaluate_batch_api_v1_transactions_evaluate_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchTransactionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Decision"
                  },
                  "type": "array",
                  "title": "Response Evaluate Batch Api V1 Transactions Evaluate Batch Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BatchTransactionInput": {
        "properties": {
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionInput"
            },
            "type": "array",
            "maxItems": 200,
            "minItems": 1,
            "title": "Transactions"
          }
        },
        "type": "object",
        "required": [
          "transactions"
        ],
        "title": "BatchTransactionInput",
        "description": "Lote para replay y modo sombra. Máximo 200 por llamada."
      },
      "Behavior": {
        "properties": {
          "tx_last_hour": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Tx Last Hour",
            "default": 0
          },
          "tx_last_24h": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Tx Last 24H",
            "default": 0
          },
          "failed_logins_24h": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Failed Logins 24H",
            "default": 0
          },
          "session_seconds": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Session Seconds",
            "default": 0
          }
        },
        "type": "object",
        "title": "Behavior"
      },
...
      "CustomerProfile": {
        "properties": {
          "segment": {
            "type": "string",
            "enum": [
              "persona_natural",
              "empresa"
            ],
            "title": "Segment",
            "default": "persona_natural"
          },
          "risk_tier": {
            "type": "string",
            "enum": [
              "bajo",
              "medio",
              "alto"
            ],
            "title": "Risk Tier",
            "default": "medio"
          }
        },
        "type": "object",
        "title": "CustomerProfile"
      },
...
      "Decision": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "process": {
            "type": "string",
            "enum": [
              "transaction",
              "document"
            ],
            "title": "Process"
          },
          "verdict": {
            "type": "string",
            "enum": [
              "aprobar",
              "validacion_adicional",
              "bloquear",
              "autentico",
              "sospechoso",
              "falso"
            ],
            "title": "Verdict"
          },
          "score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "requires_human_review": {
            "type": "boolean",
            "title": "Requires Human Review",
            "default": false
          },
          "shadow": {
            "type": "boolean",
            "title": "Shadow",
            "description": "true si SHADOW_MODE forzó el veredicto a `aprobar`.",
            "default": false
          },
          "explanation_customer": {
            "type": "string",
            "title": "Explanation Customer",
            "description": "Texto breve y amable. Jamás menciona reglas, umbrales, modelos, scores ni hallazgos forenses.",
            "default": ""
          },
          "explanation_analyst": {
            "type": "string",
            "title": "Explanation Analyst",
            "description": "Texto técnico para el analista.",
            "default": ""
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": ""
          },
          "tokens_in": {
            "type": "integer",
            "title": "Tokens In",
            "default": 0
          },
          "tokens_out": {
            "type": "integer",
            "title": "Tokens Out",
            "default": 0
          },
          "cost_usd": {
            "type": "number",
            "title": "Cost Usd",
            "default": 0.0
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "process",
          "verdict",
          "score",
          "confidence"
        ],
        "title": "Decision",
        "description": "Salida única de ambos procesos."
      },
      "DestinationAccount": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "bank": {
            "type": "string",
            "title": "Bank",
            "default": ""
          },
          "is_new_beneficiary": {
            "type": "boolean",
            "title": "Is New Beneficiary",
            "default": false
          },
          "country": {
            "type": "string",
            "title": "Country",
            "default": "CL"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "DestinationAccount"
      },
      "Device": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "default": ""
          },
          "is_new_device": {
            "type": "boolean",
            "title": "Is New Device",
            "default": false
          },
          "os": {
            "type": "string",
            "title": "Os",
            "default": ""
          },
          "ip_country": {
            "type": "string",
            "title": "Ip Country",
            "default": "CL"
          },
          "vpn": {
            "type": "boolean",
            "title": "Vpn",
            "default": false
          }
        },
        "type": "object",
        "title": "Device"
      },
...
      "Evidence": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "rule",
              "similarity",
              "vision",
              "metadata",
              "consistency"
            ],
            "title": "Type"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Identificador estable de la señal, en inglés."
          },
          "value": {
            "title": "Value",
            "description": "Valor observado (bool, número, texto o lista)."
          },
          "weight": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Weight",
            "description": "Peso relativo dentro de su fuente."
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Explicación en una frase, en español.",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "type",
          "name",
          "value",
          "weight"
        ],
        "title": "Evidence",
        "description": "Una señal que sustenta la decisión. El analista las ve ordenadas por peso."
      },
...
      "Geo": {
        "properties": {
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          },
          "distance_from_home_km": {
            "type": "number",
            "minimum": 0.0,
            "title": "Distance From Home Km",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "Geo"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OriginAccount": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "age_days": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Age Days",
            "description": "Antigüedad de la cuenta en días."
          },
          "avg_monthly_amount": {
            "type": "number",
            "minimum": 0.0,
            "title": "Avg Monthly Amount",
            "description": "Monto promedio mensual, en CLP."
          },
          "country": {
            "type": "string",
            "title": "Country",
            "default": "CL"
          }
        },
        "type": "object",
        "required": [
          "id",
          "age_days",
          "avg_monthly_amount"
        ],
        "title": "OriginAccount"
      },
...
      "TransactionInput": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "CLP"
          },
          "channel": {
            "type": "string",
            "enum": [
              "app_movil",
              "web",
              "cajero",
              "sucursal",
              "call_center",
              "api"
            ],
            "title": "Channel",
            "default": "app_movil"
          },
          "type": {
            "type": "string",
            "enum": [
              "transferencia",
              "pago",
              "retiro",
              "compra"
            ],
            "title": "Type",
            "default": "transferencia"
          },
          "origin_account": {
            "$ref": "#/components/schemas/OriginAccount"
          },
          "destination_account": {
            "$ref": "#/components/schemas/DestinationAccount"
          },
          "device": {
            "$ref": "#/components/schemas/Device"
          },
          "geo": {
            "$ref": "#/components/schemas/Geo"
          },
          "behavior": {
            "$ref": "#/components/schemas/Behavior"
          },
          "customer_profile": {
            "$ref": "#/components/schemas/CustomerProfile"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "timestamp",
          "amount",
          "origin_account",
          "destination_account"
        ],
        "title": "TransactionInput",
        "description": "Entrada del proceso A. Los identificadores son opacos (no hay PII).",
        "example": {
          "amount": 1250000,
          "behavior": {
            "failed_logins_24h": 2,
            "session_seconds": 40,
            "tx_last_24h": 7,
            "tx_last_hour": 3
          },
          "channel": "app_movil",
          "currency": "CLP",
          "customer_profile": {
            "risk_tier": "medio",
            "segment": "persona_natural"
          },
          "destination_account": {
            "bank": "OtroBanco",
            "country": "CL",
            "id": "ACC-7781",
            "is_new_beneficiary": true
          },
          "device": {
            "id": "DEV-55",
            "ip_country": "CL",
            "is_new_device": true,
            "os": "Android",
            "vpn": false
          },
          "geo": {
            "distance_from_home_km": 3.2,
            "lat": -33.45,
            "lon": -70.66
          },
          "origin_account": {
            "age_days": 1450,
            "avg_monthly_amount": 900000,
            "country": "CL",
            "id": "ACC-1001"
          },
          "timestamp": "2026-09-08T14:32:10-03:00",
          "transaction_id": "TX-000123",
          "type": "transferencia"
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}